+++ go.mod
@@ -8,13 +8,12 @@ require (
 	github.com/evanphx/json-patch v4.5.0+incompatible // indirect
 	github.com/fromanirh/numalign v0.0.2
 	github.com/ghodss/yaml v1.0.0
-	github.com/googleapis/gnostic v0.3.1 // indirect
 	github.com/hashicorp/golang-lru v0.5.4 // indirect
 	github.com/intel/sriov-network-device-plugin v0.0.0-20200603101849-e116e9c7d0b8
 	github.com/json-iterator/go v1.1.10 // indirect
 	github.com/onsi/ginkgo v1.12.1
 	github.com/onsi/gomega v1.10.1
-	github.com/opencontainers/runtime-spec v1.0.0
+	github.com/opencontainers/runtime-spec v1.0.3-0.20200520003142-237cc4f519e2
 	github.com/swatisehgal/topologyapi v0.0.0-20200802230855-6f9c5ac0d357
 	golang.org/x/text v0.3.3 // indirect
 	google.golang.org/grpc v1.28.1
@@ -29,6 +28,7 @@ require (
 
 // Pinned to kubernetes-1.18.6
 replace (
+	github.com/googleapis/gnostic => github.com/googleapis/gnostic v0.4.0
 	k8s.io/api => k8s.io/api v0.18.6
 	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.18.6
 	k8s.io/apimachinery => k8s.io/apimachinery v0.18.6
@@ -47,7 +47,7 @@ replace (
 	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.18.6
 	k8s.io/kubectl => k8s.io/kubectl v0.18.6
 	k8s.io/kubelet => k8s.io/kubelet v0.18.6
-	k8s.io/kubernetes => k8s.io/kubernetes v1.18.6
+	k8s.io/kubernetes => ../../../k8s.io/kubernetes
 	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.18.6
 	k8s.io/metrics => k8s.io/metrics v0.18.6
 	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.18.6
