+++ main.go
@@ -16,7 +16,11 @@ import (
 
 const (
 	// ProgramName is the canonical name of this program
-	ProgramName = "resource-topology-exporter"
+	ProgramName       = "resource-topology-exporter"
+	ContainerdRuntime = "containerd"
+	CRIORuntime       = "cri-o"
+	CRISource         = "cri"
+	PodResourceSource = "pod-resource-api"
 )
 
 func main() {
@@ -28,7 +32,6 @@ func main() {
 	if args.SRIOVConfigFile == "" {
 		log.Fatalf("missing SRIOV device plugin configuration file path")
 	}
-
 	klConfig, err := kubeconf.GetKubeletConfigFromLocalFile(args.KubeletConfigFile)
 	if err != nil {
 		log.Fatalf("error getting topology Manager Policy: %v", err)
@@ -44,7 +47,13 @@ func main() {
 	}
 
 	// Get new finder instance
-	instance, err := finder.NewFinder(args, pci2ResMap)
+	var finderInstance finder.Finder
+	if args.Source == CRISource {
+		finderInstance, err = finder.NewCRIFinder(args, pci2ResMap)
+	} else {
+		//args.Source == PodResourceSource
+		finderInstance, err = finder.NewPodResourceFinder(args, pci2ResMap)
+	}
 	if err != nil {
 		log.Fatalf("Failed to initialize Finder instance: %v", err)
 	}
@@ -54,14 +63,19 @@ func main() {
 		log.Fatalf("Failed to initialize crdExporter instance: %v", err)
 	}
 
+	// CAUTION: these resources are expected to change rarely - if ever. So we are intentionally do this once during the process lifecycle.
+	nodeResourceData, err := finder.NewNodeResources(args.SysfsRoot, pci2ResMap)
+	if err != nil {
+		log.Fatalf("Failed to obtain node resource information: %v", err)
+	}
 	for {
-		podResources, err := instance.Scan()
+		podResources, err := finderInstance.Scan(nodeResourceData.GetPCI2ResourceMap())
 		if err != nil {
-			log.Printf("CRI scan failed: %v\n", err)
+			log.Printf("Scan failed: %v\n", err)
 			continue
 		}
 
-		perNumaResources := instance.Aggregate(podResources)
+		perNumaResources := finder.Aggregate(podResources, nodeResourceData)
 		log.Printf("allocatedResourcesNumaInfo:%v", spew.Sdump(perNumaResources))
 
 		if err = crdExporter.CreateOrUpdate("default", perNumaResources); err != nil {
@@ -76,20 +90,21 @@ func main() {
 // The argument argv is passed only for testing purposes.
 func argsParse(argv []string) (finder.Args, error) {
 	args := finder.Args{
-		ContainerRuntime:  "containerd",
-		CRIEndpointPath:   "/host-run/containerd/containerd.sock",
+		Source:            "pod-resource-api",
 		SleepInterval:     time.Duration(3 * time.Second),
 		SysfsRoot:         "/host-sys",
 		SRIOVConfigFile:   "/etc/sriov-config/config.json",
 		KubeletConfigFile: "/host-etc/kubernetes/kubelet.conf",
 	}
 	usage := fmt.Sprintf(`Usage:
-  %s [--sleep-interval=<seconds>] [--cri-path=<path>] [--watch-namespace=<namespace>] [--sysfs=<mountpoint>] [--sriov-config-file=<path>] [--container-runtime=<runtime>] [--kubelet-config-file=<path>]
+  %s [--sleep-interval=<seconds>] [--source=<path>] [--container-runtime=<runtime>] [--cri-socket=<path>] [--podresources-socket=<path>] [--watch-namespace=<namespace>] [--sysfs=<mountpoint>] [--sriov-config-file=<path>] [--kubelet-config-file=<path>]
   %s -h | --help
   Options:
   -h --help                       Show this screen.
-  --container-runtime=<runtime>   Container Runtime to be used (containerd|cri-o). [Default: %v]
-  --cri-path=<path>               CRI Endpoint file path to use. [Default: %v]
+	--source=<source>								Evaluation source to be used (pod-resource-api|cri). [Default: %v]
+  --container-runtime=<runtime>   Container Runtime to be used (containerd|cri-o).
+  --cri-socket=<path>             CRI Socket path to use.
+	--podresources-socket=<path>    Pod Resource Socket path to use.
   --sleep-interval=<seconds>      Time to sleep between updates. [Default: %v]
   --watch-namespace=<namespace>   Namespace to watch pods for. Use "" for all namespaces.
   --sysfs=<mountpoint>            Mount point of the sysfs. [Default: %v]
@@ -97,8 +112,7 @@ func argsParse(argv []string) (finder.Args, error) {
   --kubelet-config-file=<path>    Kubelet config file path. [Default: %v]`,
 		ProgramName,
 		ProgramName,
-		args.ContainerRuntime,
-		args.CRIEndpointPath,
+		args.Source,
 		args.SleepInterval,
 		args.SysfsRoot,
 		args.SRIOVConfigFile,
@@ -118,12 +132,42 @@ func argsParse(argv []string) (finder.Args, error) {
 		args.KubeletConfigFile = kubeletConfigPath
 	}
 	args.SysfsRoot = arguments["--sysfs"].(string)
-	runtime := arguments["--container-runtime"].(string)
-	if !(runtime == "containerd" || runtime == "cri-o") {
-		return args, fmt.Errorf("invalid --container-runtime specified")
+
+	if source, ok := arguments["--source"].(string); ok {
+		args.Source = source
 	}
-	args.ContainerRuntime = runtime
-	args.CRIEndpointPath = arguments["--cri-path"].(string)
+	if args.Source != PodResourceSource && args.Source != CRISource {
+		return args, fmt.Errorf("invalid --source specified")
+
+	} else if args.Source == PodResourceSource {
+		//podresource source
+		if path, ok := arguments["--podresources-socket"].(string); ok {
+			args.PodResourceSocketPath = path
+		}
+		//return error in case cri-socket path is specified in case of pod-resource-socket source
+		if _, ok := arguments["--cri-socket"].(string); ok {
+			return args, fmt.Errorf("No need to specify CRI socket path in case pod-resource-api is specified as the source")
+		}
+		//return error in case container-runtime is specified in case of pod-resource-socket source
+		if _, ok := arguments["--container-runtime"].(string); ok {
+			return args, fmt.Errorf("No need to specify container runtime in case pod-resource-api is specified as the source")
+		}
+	} else {
+		//cri source
+		if path, ok := arguments["--cri-socket"].(string); ok {
+			args.CRISocketPath = path
+		}
+		runtime := arguments["--container-runtime"].(string)
+		if runtime != ContainerdRuntime && runtime != CRIORuntime {
+			return args, fmt.Errorf("invalid --container-runtime specified")
+		}
+		args.ContainerRuntime = runtime
+		//return error in case pod-resource-socket path is specified in case of cri source
+		if _, ok := arguments["--podresources-socket"].(string); ok {
+			return args, fmt.Errorf("No need to specify Pod Resource socket path in case CRI is specified as the source")
+		}
+	}
+
 	args.SleepInterval, err = time.ParseDuration(arguments["--sleep-interval"].(string))
 	if err != nil {
 		return args, fmt.Errorf("invalid --sleep-interval specified: %s", err.Error())
